consumer tool optimizations
In case of bulk-consume retry in consumer tool, the tool store  'X-CMWELL-TO' header
in order to send it on retry. This way it should ease the stress on the
server.

ConsumerMain now sends requests to _out only when needed (i.e., does not
send requests to _out when format is text or tsv.